Preview Feature not working
I have Enen no Shouboutai from 2019 in my shelf, and there is a new season in 2025. Its also showing me this in seasonal with Add all to shelf, but in previews it is not showing anything. In Preview should be 2 sections.

First should be future Seasonal Anime and in the second section should be continuations regadless of the current season. Example: i have an anime from 2019 in my shelf and there is a continuation from 2022 and one from 2026. 2022 should be in second section and 2026 should be in first section