My Shelf Grouping
On My Shelf animes from the same anime, so an anime and its continuations, should be grouped together. The Pictures should be stacked on top of each other and a counter should be visibile on the left corner. The Status should be displayed as the lowest of them all, the rating should be the middle of them all, and the episode count should be added together. The grouping cannot be edited, but when clicked on the grouping a modal should be opend with all the anime continuations and their individual settings, which can be edited