Import and Export baseline
There Should also be an Import and Export function as an Icon. When clicked on it, it should show a Modal where you can select Import or export.

When Exporting, all Data from My Shelf should be exported as an csv file

When Importing There should be multiple possibilities, but for now only csv import. When selecting csv import it should show the rownames its expecting and also show us an file dialog. All Data should be Uploaded to My Shelf. After Importing it should show as an result which animes could be imported and which not